requirements.txt
py-clob-client>=0.34.0
httpx>=0.24.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
web3>=6.0.0
psycopg2-binary>=2.9.0
